LayoutTests/http/wpt/webauthn/public-key-credential-get-success-u2f.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[get]] with minimum options in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with more allow credentials in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with test of user presence in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with empty extensions in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with same site AppID but not used in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with empty AppID in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with an AppID in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with multiple credentials and AppID is not used in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with multiple credentials and AppID is used in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with downgraded authenticator in a mock hid authenticator.
PASS PublicKeyCredential's [[get]] with downgraded authenticator in a mock hid authenticator. (AppID)
